refactor, headless client now runs on Windows
