+++ rust/headless-client/src/imp_linux.rs
@@ -2,7 +2,6 @@
 
 use super::{Cli, Cmd, IpcClientMsg, IpcServerMsg, TOKEN_ENV_KEY};
 use anyhow::{bail, Context, Result};
-use clap::Parser;
 use connlib_client_shared::{file_logger, Callbacks, ResourceDescription, Session, Sockets};
 use connlib_shared::{
     keypair,
@@ -23,7 +22,7 @@ use std::{
 };
 use tokio::{
     net::{UnixListener, UnixStream},
-    signal::unix::SignalKind,
+    signal::unix::SignalKind as TokioSignalKind,
     sync::mpsc,
 };
 use tokio_util::codec::{FramedRead, FramedWrite, LengthDelimitedCodec};
@@ -46,94 +45,43 @@ const MAX_PARTITION_TIME: Duration = Duration::from_secs(60 * 60 * 24 * 30);
 /// on some systems, `/run` should be the newer version.
 pub const SOCK_PATH: &str = "/run/firezone-client.sock";
 
+pub(crate) struct Signals {
+    sighup: tokio::signal::unix::Signal,
+    sigint: tokio::signal::unix::Signal,
+}
+
+impl Signals {
+    pub(crate) fn new() -> Result<Self> {
+        let sighup = tokio::signal::unix::signal(TokioSignalKind::hangup())?;
+        let sigint = tokio::signal::unix::signal(TokioSignalKind::interrupt())?;
+
+        Ok(Self {
+            sighup,
+            sigint,
+        })
+    }
+
+    pub(crate) fn poll(&mut self, cx: &mut Context) -> Poll<super::SignalKind> {
+        if self.sigint.poll_recv(cx).is_ready() {
+            return Poll::Ready(super::SignalKind::Interrupt);
+        }
+
+        if self.sighup.poll_recv(cx).is_ready() {
+            return Poll::Ready(super::SignalKind::Hangup)
+        }
+
+        Poll::Pending
+    }
+}
+
 pub fn default_token_path() -> PathBuf {
     PathBuf::from("/etc")
         .join(connlib_shared::BUNDLE_ID)
         .join("token")
 }
 
-pub fn run() -> Result<()> {
-    let mut cli = Cli::parse();
-
-    // Modifying the environment of a running process is unsafe. If any other
-    // thread is reading or writing the environment, something bad can happen.
-    // So `run` must take over as early as possible during startup, and
-    // take the token env var before any other threads spawn.
-
-    let token_env_var = cli.token.take().map(SecretString::from);
-    let cli = cli;
-
-    // Docs indicate that `remove_var` should actually be marked unsafe
-    // SAFETY: We haven't spawned any other threads, this code should be the first
-    // thing to run after entering `main`. So nobody else is reading the environment.
-    #[allow(unused_unsafe)]
-    unsafe {
-        // This removes the token from the environment per <https://security.stackexchange.com/a/271285>. We run as root so it may not do anything besides defense-in-depth.
-        std::env::remove_var(TOKEN_ENV_KEY);
-    }
-    assert!(std::env::var(TOKEN_ENV_KEY).is_err());
-
-    let (layer, _handle) = cli.log_dir.as_deref().map(file_logger::layer).unzip();
-    setup_global_subscriber(layer);
-
-    tracing::info!(git_version = crate::GIT_VERSION);
-
-    match cli.command() {
-        Cmd::Auto => {
-            if let Some(token) = get_token(token_env_var, &cli)? {
-                run_standalone(cli, &token)
-            } else {
-                run_ipc_service(cli)
-            }
-        }
-        Cmd::IpcService => run_ipc_service(cli),
-        Cmd::Standalone => {
-            let token = get_token(token_env_var, &cli)?.with_context(|| {
-                format!(
-                    "Can't find the Firezone token in ${TOKEN_ENV_KEY} or in `{}`",
-                    cli.token_path
-                )
-            })?;
-            run_standalone(cli, &token)
-        }
-        Cmd::StubIpcClient => run_debug_ipc_client(cli),
-    }
-}
-
-/// Read the token from disk if it was not in the environment
-///
-/// # Returns
-/// - `Ok(None)` if there is no token to be found
-/// - `Ok(Some(_))` if we found the token
-/// - `Err(_)` if we found the token on disk but failed to read it
-fn get_token(token_env_var: Option<SecretString>, cli: &Cli) -> Result<Option<SecretString>> {
-    // This is very simple but I don't want to write it twice
-    if let Some(token) = token_env_var {
-        return Ok(Some(token));
-    }
-    read_token_file(cli)
-}
-
-/// Try to retrieve the token from disk
-///
-/// Sync because we do blocking file I/O
-fn read_token_file(cli: &Cli) -> Result<Option<SecretString>> {
-    let path = PathBuf::from(&cli.token_path);
-
-    if let Ok(token) = std::env::var(TOKEN_ENV_KEY) {
-        std::env::remove_var(TOKEN_ENV_KEY);
-
-        let token = SecretString::from(token);
-        // Token was provided in env var
-        tracing::info!(
-            ?path,
-            ?TOKEN_ENV_KEY,
-            "Found token in env var, ignoring any token that may be on disk."
-        );
-        return Ok(Some(token));
-    }
-
-    let Ok(stat) = nix::sys::stat::fstatat(None, &path, nix::fcntl::AtFlags::empty()) else {
+pub(crate) fn check_token_permissions(path: &Path) -> Result<()> {
+    let Ok(stat) = nix::sys::stat::fstatat(None, path, nix::fcntl::AtFlags::empty()) else {
         // File doesn't exist or can't be read
         tracing::info!(
             ?path,
@@ -160,107 +108,15 @@ fn read_token_file(cli: &Cli) -> Result<Option<SecretString>> {
             path.display()
         );
     }
-
-    let Ok(bytes) = std::fs::read(&path) else {
-        // We got the metadata a second ago, but can't read the file itself.
-        // Pretty strange, would have to be a disk fault or TOCTOU.
-        tracing::info!(?path, "Token file existed but now is unreadable");
-        return Ok(None);
-    };
-    let token = String::from_utf8(bytes)?.trim().to_string();
-    let token = SecretString::from(token);
-
-    tracing::info!(?path, "Loaded token from disk");
-    Ok(Some(token))
+    Ok(())
 }
 
-fn run_standalone(cli: Cli, token: &SecretString) -> Result<()> {
-    tracing::info!("Running in standalone mode");
-    let rt = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()?;
-    let _guard = rt.enter();
-    let max_partition_time = cli.max_partition_time.map(|d| d.into());
-
-    // AKA "Device ID", not the Firezone slug
-    let firezone_id = match cli.firezone_id {
-        Some(id) => id,
-        None => connlib_shared::device_id::get().context("Could not get `firezone_id` from CLI, could not read it from disk, could not generate it and save it to disk")?.id,
-    };
-
-    let (private_key, public_key) = keypair();
-    let login = LoginUrl::client(cli.api_url, token, firezone_id, None, public_key.to_bytes())?;
-
-    if cli.check {
-        tracing::info!("Check passed");
-        return Ok(());
-    }
-
-    let session = Session::connect(
-        login,
-        Sockets::new(),
-        private_key,
-        None,
-        CallbackHandler,
-        max_partition_time,
-        rt.handle().clone(),
-    );
-    // TODO: this should be added dynamically
-    session.set_dns(system_resolvers(get_dns_control_from_env()).unwrap_or_default());
-
-    let mut sigint = tokio::signal::unix::signal(SignalKind::interrupt())?;
-    let mut sighup = tokio::signal::unix::signal(SignalKind::hangup())?;
-
-    let result = rt.block_on(async {
-        future::poll_fn(|cx| loop {
-            if sigint.poll_recv(cx).is_ready() {
-                tracing::debug!("Received SIGINT");
-
-                return Poll::Ready(std::io::Result::Ok(()));
-            }
-
-            if sighup.poll_recv(cx).is_ready() {
-                tracing::debug!("Received SIGHUP");
-
-                session.reconnect();
-                continue;
-            }
-
-            return Poll::Pending;
-        })
-        .await
-    });
-
-    session.disconnect();
-
-    Ok(result?)
-}
-
-fn system_resolvers(dns_control_method: Option<DnsControlMethod>) -> Result<Vec<IpAddr>> {
-    match dns_control_method {
+pub(crate) fn system_resolvers() -> Result<Vec<IpAddr>> {
+    match get_dns_control_from_env() {
         None => get_system_default_resolvers_resolv_conf(),
         Some(DnsControlMethod::EtcResolvConf) => get_system_default_resolvers_resolv_conf(),
         Some(DnsControlMethod::NetworkManager) => get_system_default_resolvers_network_manager(),
         Some(DnsControlMethod::Systemd) => get_system_default_resolvers_systemd_resolved(),
-    }
-}
-
-#[derive(Clone)]
-struct CallbackHandler;
-
-impl Callbacks for CallbackHandler {
-    fn on_disconnect(&self, error: &connlib_client_shared::Error) {
-        tracing::error!("Disconnected: {error}");
-
-        std::process::exit(1);
-    }
-
-    fn on_update_resources(&self, resources: Vec<connlib_client_shared::ResourceDescription>) {
-        // See easily with `export RUST_LOG=firezone_headless_client=debug`
-        tracing::debug!(len = resources.len(), "Printing the resource list one time");
-        for resource in &resources {
-            tracing::debug!(?resource);
-        }
     }
 }
 
@@ -314,22 +170,7 @@ fn parse_resolvectl_output(s: &str) -> Vec<IpAddr> {
         .collect()
 }
 
-fn run_debug_ipc_client(_cli: Cli) -> Result<()> {
-    let rt = tokio::runtime::Runtime::new()?;
-    rt.block_on(async {
-        tracing::info!(pid = std::process::id(), "run_debug_ipc_client");
-        let stream = UnixStream::connect(SOCK_PATH)
-            .await
-            .with_context(|| format!("couldn't connect to UDS at {SOCK_PATH}"))?;
-        let mut stream = IpcStream::new(stream, LengthDelimitedCodec::new());
-
-        stream.send(serde_json::to_string("Hello")?.into()).await?;
-        Ok::<_, anyhow::Error>(())
-    })?;
-    Ok(())
-}
-
-fn run_ipc_service(_cli: Cli) -> Result<()> {
+pub(crate) fn run_ipc_service(_cli: Cli) -> Result<()> {
     let rt = tokio::runtime::Runtime::new()?;
     tracing::info!("run_daemon");
     rt.block_on(async { ipc_listen().await })
