+++ rust/headless-client/src/lib.rs
@@ -8,10 +8,14 @@
 //! Tauri deb bundler to pick it up easily.
 //! Otherwise we would just make it a normal binary crate.
 
-use connlib_client_shared::ResourceDescription;
-use std::{net::IpAddr, path::PathBuf};
+use anyhow::{Context, Result};
+use clap::Parser;
+use connlib_client_shared::{Callbacks, file_logger, keypair, LoginUrl, ResourceDescription, Session, Sockets};
+use firezone_cli_utils::setup_global_subscriber;
+use secrecy::SecretString;
+use std::{future, net::IpAddr, path::PathBuf, task::Poll};
 
-pub use imp::{default_token_path, run};
+use imp::default_token_path;
 
 #[cfg(target_os = "linux")]
 pub mod imp_linux;
@@ -19,25 +23,9 @@ pub mod imp_linux;
 pub use imp_linux as imp;
 
 #[cfg(target_os = "windows")]
-mod imp_windows {
-    use clap::Parser;
-
-    pub fn default_token_path() -> std::path::PathBuf {
-        todo!()
-    }
-
-    pub fn run() -> anyhow::Result<()> {
-        let cli = super::Cli::parse();
-        let _cmd = cli.command();
-        tracing::info!(git_version = crate::GIT_VERSION);
-        // Clippy will complain that the `Result` type is pointless if we can't
-        // possibly throw an error, because it doesn't see that the Linux impl does
-        // throw errors
-        anyhow::bail!("`headless-client` is not implemented for Windows yet");
-    }
-}
+pub mod imp_windows;
 #[cfg(target_os = "windows")]
-use imp_windows as imp;
+pub use imp_windows as imp;
 
 /// Output of `git describe` at compile time
 /// e.g. `1.0.0-pre.4-20-ged5437c88-modified` where:
@@ -125,9 +113,6 @@ enum Cmd {
     IpcService,
     /// Act as a CLI-only Client
     Standalone,
-    /// Act as an IPC client for development
-    #[command(hide = true)]
-    StubIpcClient,
 }
 
 #[derive(Debug, serde::Deserialize, serde::Serialize)]
@@ -144,4 +129,180 @@ pub enum IpcServerMsg {
     OnDisconnect,
     OnUpdateResources(Vec<ResourceDescription>),
     TunnelReady,
+}
+
+pub fn run() -> Result<()> {
+    let mut cli = Cli::parse();
+
+    // Modifying the environment of a running process is unsafe. If any other
+    // thread is reading or writing the environment, something bad can happen.
+    // So `run` must take over as early as possible during startup, and
+    // take the token env var before any other threads spawn.
+
+    let token_env_var = cli.token.take().map(SecretString::from);
+    let cli = cli;
+
+    // Docs indicate that `remove_var` should actually be marked unsafe
+    // SAFETY: We haven't spawned any other threads, this code should be the first
+    // thing to run after entering `main`. So nobody else is reading the environment.
+    #[allow(unused_unsafe)]
+    unsafe {
+        // This removes the token from the environment per <https://security.stackexchange.com/a/271285>. We run as root so it may not do anything besides defense-in-depth.
+        std::env::remove_var(TOKEN_ENV_KEY);
+    }
+    assert!(std::env::var(TOKEN_ENV_KEY).is_err());
+
+    let (layer, _handle) = cli.log_dir.as_deref().map(file_logger::layer).unzip();
+    setup_global_subscriber(layer);
+
+    tracing::info!(git_version = crate::GIT_VERSION);
+
+    match cli.command() {
+        Cmd::Auto => {
+            if let Some(token) = get_token(token_env_var, &cli)? {
+                run_standalone(cli, &token)
+            } else {
+                imp::run_ipc_service(cli)
+            }
+        }
+        Cmd::IpcService => imp::run_ipc_service(cli),
+        Cmd::Standalone => {
+            let token = get_token(token_env_var, &cli)?.with_context(|| {
+                format!(
+                    "Can't find the Firezone token in ${TOKEN_ENV_KEY} or in `{}`",
+                    cli.token_path
+                )
+            })?;
+            run_standalone(cli, &token)
+        }
+    }
+}
+
+// Allow dead code because Windows doesn't have an obvious SIGHUP equivalent
+#[allow(dead_code)]
+enum SignalKind {
+    Hangup,
+    Interrupt,
+}
+
+fn run_standalone(cli: Cli, token: &SecretString) -> Result<()> {
+    tracing::info!("Running in standalone mode");
+    let rt = tokio::runtime::Builder::new_current_thread()
+        .enable_all()
+        .build()?;
+    let _guard = rt.enter();
+    let max_partition_time = cli.max_partition_time.map(|d| d.into());
+
+    // AKA "Device ID", not the Firezone slug
+    let firezone_id = match cli.firezone_id {
+        Some(id) => id,
+        None => connlib_shared::device_id::get().context("Could not get `firezone_id` from CLI, could not read it from disk, could not generate it and save it to disk")?.id,
+    };
+
+    let (private_key, public_key) = keypair();
+    let login = LoginUrl::client(cli.api_url, token, firezone_id, None, public_key.to_bytes())?;
+
+    if cli.check {
+        tracing::info!("Check passed");
+        return Ok(());
+    }
+
+    let session = Session::connect(
+        login,
+        Sockets::new(),
+        private_key,
+        None,
+        CallbackHandler,
+        max_partition_time,
+        rt.handle().clone(),
+    );
+    // TODO: this should be added dynamically
+    session.set_dns(imp::system_resolvers().unwrap_or_default());
+
+    let mut signals = imp::Signals::new()?;
+
+    let result = rt.block_on(async {
+        future::poll_fn(|cx| loop {
+            match signals.poll(cx) {
+                Poll::Ready(SignalKind::Hangup) => {
+                    session.reconnect();
+                    continue;
+                }
+                Poll::Ready(SignalKind::Interrupt) => return Poll::Ready(std::io::Result::Ok(())),
+                Poll::Pending => return Poll::Pending,
+            }
+        })
+        .await
+    });
+
+    session.disconnect();
+
+    Ok(result?)
+}
+
+#[derive(Clone)]
+struct CallbackHandler;
+
+impl Callbacks for CallbackHandler {
+    fn on_disconnect(&self, error: &connlib_client_shared::Error) {
+        tracing::error!("Disconnected: {error}");
+
+        std::process::exit(1);
+    }
+
+    fn on_update_resources(&self, resources: Vec<connlib_client_shared::ResourceDescription>) {
+        // See easily with `export RUST_LOG=firezone_headless_client=debug`
+        tracing::debug!(len = resources.len(), "Printing the resource list one time");
+        for resource in &resources {
+            tracing::debug!(?resource);
+        }
+    }
+}
+
+/// Read the token from disk if it was not in the environment
+///
+/// # Returns
+/// - `Ok(None)` if there is no token to be found
+/// - `Ok(Some(_))` if we found the token
+/// - `Err(_)` if we found the token on disk but failed to read it
+fn get_token(token_env_var: Option<SecretString>, cli: &Cli) -> Result<Option<SecretString>> {
+    // This is very simple but I don't want to write it twice
+    if let Some(token) = token_env_var {
+        return Ok(Some(token));
+    }
+    read_token_file(cli)
+}
+
+/// Try to retrieve the token from disk
+///
+/// Sync because we do blocking file I/O
+fn read_token_file(cli: &Cli) -> Result<Option<SecretString>> {
+    let path = PathBuf::from(&cli.token_path);
+
+    if let Ok(token) = std::env::var(TOKEN_ENV_KEY) {
+        std::env::remove_var(TOKEN_ENV_KEY);
+
+        let token = SecretString::from(token);
+        // Token was provided in env var
+        tracing::info!(
+            ?path,
+            ?TOKEN_ENV_KEY,
+            "Found token in env var, ignoring any token that may be on disk."
+        );
+        return Ok(Some(token));
+    }
+
+    imp::check_token_permissions(&path)?;
+
+    let Ok(bytes) = std::fs::read(&path) else {
+        // We got the metadata a second ago, but can't read the file itself.
+        // Pretty strange, would have to be a disk fault or TOCTOU.
+        tracing::info!(?path, "Token file existed but now is unreadable");
+        return Ok(None);
+    };
+    let token = String::from_utf8(bytes)?.trim().to_string();
+    let token = SecretString::from(token);
+
+    tracing::info!(?path, "Loaded token from disk");
+    Ok(Some(token))
 }
